refactor: split `http` module
In preparation of upcoming changes, I would like to split this module
into smaller files. No functional change, just moves code around.
